@@ -1,7 +1,7 @@
 {
   "name": "phaser-npm-webpack-typescript-starter-project",
   "productName": "phaser-npm-webpack-typescript-starter-project",
-  "version": "1.7.3",
+  "version": "1.7.8",
   "main": "electron-main.js",
   "scripts": {
     "webpack:dev": "webpack --config webpack.dev.config.js --progress --colors",
@@ -30,27 +30,28 @@
   },
   "homepage": "https://github.com/rroylance/phaser-npm-webpack-typescript-starter-project#readme",
   "devDependencies": {
-    "@types/node": "7.0.5",
-    "clean-webpack-plugin": "0.1.15",
+    "@types/node": "^7.0.11",
+    "clean-webpack-plugin": "0.1.16",
     "commander": "2.9.0",
     "copy-webpack-plugin": "4.0.1",
     "echo-cli": "1.0.8",
-    "electron": "1.4.15",
-    "electron-packager": "8.5.2",
+    "electron": "1.6.2",
+    "electron-packager": "8.6.0",
+    "expose-loader": "0.7.3",
     "file-loader": "0.10.1",
     "git-revision-webpack-plugin": "2.4.1",
     "html-webpack-plugin": "2.28.0",
     "npm-check": "5.4.0",
-    "phaser-ce": "2.7.3",
+    "phaser-ce": "2.7.5",
     "pngjs-image": "0.11.7",
-    "shelljs": "0.7.6",
-    "ts-loader": "2.0.1",
-    "tslint": "4.5.0",
-    "tslint-loader": "3.4.2",
+    "shelljs": "0.7.7",
+    "ts-loader": "2.0.3",
+    "tslint": "4.5.1",
+    "tslint-loader": "3.4.3",
     "typescript": "2.2.1",
     "webfontloader": "1.6.27",
-    "webpack": "2.2.1",
-    "webpack-dev-server": "2.4.1",
+    "webpack": "^2.3.2",
+    "webpack-dev-server": "2.4.2",
     "webpack-shell-plugin": "0.5.0",
     "xml2js": "0.4.17"
   }
