@@ -1,6 +1,6 @@
-(global as any).p2 = require('p2');
-(global as any).PIXI = require('pixi');
-(global as any).Phaser = require('phaser');
+import 'p2';
+import 'pixi';
+import 'phaser';
 
 import * as WebFontLoader from 'webfontloader';
 
@@ -19,7 +19,6 @@ class App extends Phaser.Game {
         this.state.add('preloader', Preloader);
         // this.state.add('title', Title);
         this.state.add('pong', Pong);
-
         this.state.start('boot');
     }
 }
