@@ -40,6 +40,7 @@ var gameAssets = {};
 
 var loaderTypes = {
     image: {},
+    spritesheet: {},
     atlas: {},
     audio: {},
     audiosprite: {},
@@ -48,6 +49,8 @@ var loaderTypes = {
     json: {},
     xml: {},
     text: {},
+    script: {},
+    shader: {},
     misc: {}
 };
 
@@ -58,12 +61,14 @@ var bitmapFontExtensions = ['xml', 'fnt'];
 var jsonExtensions = ['json'];
 var xmlExtensions = ['xml'];
 var textExtensions = ['txt'];
+var scriptExtensions = ['js'];
+var shaderExtensions = ['frag'];
 
 shell.ls('assets/**/*.*').forEach(function (file) {
     var filePath = file.replace('assets/', '').split('.');
 
     gameAssets[filePath[0]] = gameAssets[filePath[0]] || [];
-    gameAssets[filePath[0]].push(filePath[1]);
+    gameAssets[filePath[0]] = gameAssets[filePath[0]].concat(filePath.slice(1));
 });
 
 for (var i in gameAssets) {
@@ -74,6 +79,8 @@ for (var i in gameAssets) {
     var jsonType = findExtension(gameAssets[i], jsonExtensions);
     var xmlType = findExtension(gameAssets[i], xmlExtensions);
     var textType = findExtension(gameAssets[i], textExtensions);
+    var scriptType = findExtension(gameAssets[i], scriptExtensions);
+    var shaderType = findExtension(gameAssets[i], shaderExtensions);
 
     if (bitmapFontType) {
         var isItActuallyAFont = false;
@@ -100,7 +107,12 @@ for (var i in gameAssets) {
         if (jsonType || xmlType) {
             loaderTypes.atlas[i] = gameAssets[i];
         } else {
-            loaderTypes.image[i] = gameAssets[i];
+            var spritesheetData = gameAssets[i][0].match(/\[(-?[0-9],?)*]/);
+            if (spritesheetData && spritesheetData.length > 0) {
+                loaderTypes.spritesheet[i] = gameAssets[i];
+            } else {
+                loaderTypes.image[i] = gameAssets[i];
+            }
         }
     } else if (fontType) {
         loaderTypes.font[i] = gameAssets[i];
@@ -110,6 +122,10 @@ for (var i in gameAssets) {
         loaderTypes.xml[i] = gameAssets[i];
     } else if (textType) {
         loaderTypes.text[i] = gameAssets[i];
+    } else if (scriptType) {
+        loaderTypes.script[i] = gameAssets[i];
+    }  else if (shaderType) {
+        loaderTypes.shader[i] = gameAssets[i];
     } else {
         loaderTypes.misc[i] = gameAssets[i];
     }
@@ -131,6 +147,32 @@ if (!Object.keys(loaderTypes.image).length) {
         for (var t in loaderTypes.image[i]) {
             shell.ShellString('\n        static get' + loaderTypes.image[i][t].toUpperCase() + '(): string { return require(\'assets/' + i + '.' + loaderTypes.image[i][t] + '\'); };').toEnd(assetsClassFile);
         }
+
+        shell.ShellString('\n    }').toEnd(assetsClassFile);
+    }
+}
+shell.ShellString('\n}\n\n').toEnd(assetsClassFile);
+
+shell.ShellString('export namespace Spritesheets {').toEnd(assetsClassFile);
+if (!Object.keys(loaderTypes.spritesheet).length) {
+    shell.ShellString('\n    class IExistSoTypeScriptWillNotComplainAboutAnEmptyNamespace {}').toEnd(assetsClassFile);
+} else {
+    for (var i in loaderTypes.spritesheet) {
+        shell.ShellString('\n    export class ' + toPascalCase(i) + ' {').toEnd(assetsClassFile);
+        shell.ShellString('\n        static getName(): string { return \'' + i.split('/').pop() + '\'; };\n').toEnd(assetsClassFile);
+
+        shell.ShellString('\n        static get' + loaderTypes.spritesheet[i][1].toUpperCase() + '(): string { return require(\'assets/' + i + '.' + loaderTypes.spritesheet[i][0] + '.' + loaderTypes.spritesheet[i][1] + '\'); };').toEnd(assetsClassFile);
+
+        var spritesheetProperties = loaderTypes.spritesheet[i][0].replace('[', '').replace(']', '').split(',');
+        if (spritesheetProperties.length < 2 || spritesheetProperties.length > 5) {
+            console.log('Invalid number of Spritesheet properties provided for \'' + i + '\'. Must have between 2 and 5; [frameWidth, frameHeight, frameMax, margin, spacing] frameWidth and frameHeight are required');
+        }
+
+        shell.ShellString('\n        static getFrameWidth(): number { return ' + parseInt(spritesheetProperties[0] ? spritesheetProperties[0] : -1) + '; };').toEnd(assetsClassFile);
+        shell.ShellString('\n        static getFrameHeight(): number { return ' + parseInt(spritesheetProperties[1] ? spritesheetProperties[1] : -1) + '; };').toEnd(assetsClassFile);
+        shell.ShellString('\n        static getFrameMax(): number { return ' + parseInt(spritesheetProperties[2] ? spritesheetProperties[2] : -1) + '; };').toEnd(assetsClassFile);
+        shell.ShellString('\n        static getMargin(): number { return ' + parseInt(spritesheetProperties[3] ? spritesheetProperties[3] : 0) + '; };').toEnd(assetsClassFile);
+        shell.ShellString('\n        static getSpacing(): number { return ' + parseInt(spritesheetProperties[4] ? spritesheetProperties[4] : 0) + '; };').toEnd(assetsClassFile);
 
         shell.ShellString('\n    }').toEnd(assetsClassFile);
     }
@@ -379,6 +421,40 @@ if (!Object.keys(loaderTypes.text).length) {
     }
 }
 shell.ShellString('\n}\n\n').toEnd(assetsClassFile);
+
+shell.ShellString('export namespace Scripts {').toEnd(assetsClassFile);
+if (!Object.keys(loaderTypes.script).length) {
+    shell.ShellString('\n    class IExistSoTypeScriptWillNotComplainAboutAnEmptyNamespace {}').toEnd(assetsClassFile);
+} else {
+    for (var i in loaderTypes.script) {
+        shell.ShellString('\n    export class ' + toPascalCase(i) + ' {').toEnd(assetsClassFile);
+        shell.ShellString('\n        static getName(): string { return \'' + i.split('/').pop() + '\'; };\n').toEnd(assetsClassFile);
+
+        for (var t in loaderTypes.script[i]) {
+            shell.ShellString('\n        static get' + loaderTypes.script[i][t].toUpperCase() + '(): string { return require(\'assets/' + i + '.' + loaderTypes.script[i][t] + '\'); };').toEnd(assetsClassFile);
+        }
+
+        shell.ShellString('\n    }').toEnd(assetsClassFile);
+    }
+}
+shell.ShellString('\n}\n').toEnd(assetsClassFile);
+
+shell.ShellString('export namespace Shaders {').toEnd(assetsClassFile);
+if (!Object.keys(loaderTypes.shader).length) {
+    shell.ShellString('\n    class IExistSoTypeScriptWillNotComplainAboutAnEmptyNamespace {}').toEnd(assetsClassFile);
+} else {
+    for (var i in loaderTypes.shader) {
+        shell.ShellString('\n    export class ' + toPascalCase(i) + ' {').toEnd(assetsClassFile);
+        shell.ShellString('\n        static getName(): string { return \'' + i.split('/').pop() + '\'; };\n').toEnd(assetsClassFile);
+
+        for (var t in loaderTypes.shader[i]) {
+            shell.ShellString('\n        static get' + loaderTypes.shader[i][t].toUpperCase() + '(): string { return require(\'assets/' + i + '.' + loaderTypes.shader[i][t] + '\'); };').toEnd(assetsClassFile);
+        }
+
+        shell.ShellString('\n    }').toEnd(assetsClassFile);
+    }
+}
+shell.ShellString('\n}\n').toEnd(assetsClassFile);
 
 shell.ShellString('export namespace Misc {').toEnd(assetsClassFile);
 if (!Object.keys(loaderTypes.misc).length) {
