@@ -28,7 +28,7 @@ export default class Preloader extends Phaser.State {
 
     private startGame(): void {
         this.game.camera.onFadeComplete.addOnce(this.loadTitle, this);
-        this.game.camera.fade(0x000000, 500);
+        this.game.camera.fade(0x000000, 1000);
     }
 
     private loadTitle(): void {
